+++ test/TestAppIsolated/ToolFunctions.cs
@@ -1,4 +1,5 @@
 using System.ComponentModel;
+using System.Globalization;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Extensions.Mcp;
 using Microsoft.Extensions.Logging;
@@ -31,6 +32,16 @@ public class TestFunction
             Hello, {entityToGreet}! Job: {job ?? "Unemployed"} | Age: {age} | Happy? {isHappy}.
             Attributes: {(attributes.Any() ? string.Join(", ", attributes) : "(none)")}
             """;
+    }
+
+    [Function(nameof(BirthdayTracker))]
+    public string BirthdayTracker(
+        [McpToolTrigger(nameof(BirthdayTracker), "Capture user birthday information.")] ToolInvocationContext context,
+        [McpToolProperty(nameof(userId), "User ID")] Guid userId,
+        [McpToolProperty(nameof(birthday), "Birthday")] DateTime birthday)
+    {
+        var date = birthday.ToUniversalTime().ToString("o", CultureInfo.InvariantCulture);
+        return $"Hello {userId}, {date}";
     }
 
     [Function(nameof(SingleArgumentFunction))]